recording builder owned by sph simulation
