+++ sphinxsim/sph_simulation/fluid_simulation_builder.cpp
@@ -15,6 +15,7 @@ void FluidSimulationBuilder::buildSimulation(SPHSimulation &sim, const json &con
     //----------------------------------------------------------------------
     SPHSystem &sph_system = sim.defineSPHSystem();
     EntityManager &config_manager = sim.getConfigManager();
+    RecordingBuilder &recording_builder = sim.getRecordingBuilder();
     ScalingConfig &scaling_config = config_manager.getEntity<ScalingConfig>("ScalingConfig");
     //----------------------------------------------------------------------
     // Creating bodies with inital geometry, materials and particles.
@@ -79,7 +80,7 @@ void FluidSimulationBuilder::buildSimulation(SPHSimulation &sim, const json &con
     //----------------------------------------------------------------------
     // Define basic state recording for visualization the simulation results.
     //----------------------------------------------------------------------
-    auto &body_state_recorder = recording_builder_ptr_->createBodyStatesRecording(
+    auto &body_state_recorder = recording_builder.createBodyStatesRecording(
         sph_system, config_manager, main_methods, config);
     //----------------------------------------------------------------------
     //	Define time integration method, screen out uput and observation sample rate.
@@ -174,7 +175,7 @@ void FluidSimulationBuilder::buildSimulation(SPHSimulation &sim, const json &con
     //----------------------------------------------------------------------
     // Define optional methods using hooking point in stage pipelines.
     //----------------------------------------------------------------------
-    recording_builder_ptr_->buildObservationIfPresent(sim, main_methods, config);
+    recording_builder.buildObservationIfPresent(sim, main_methods, config);
     buildExternalForceIfPresent(sim, main_methods, fluid_body, config);
     buildSurfaceIndicationIfOpenBoundary(sim, main_methods, fluid_inner, fluid_wall_contact);
     buildTransportVelocityFormulationIfNotFreeSurface(sim, main_methods, fluid_inner, fluid_wall_contact);
